BarterAI/tests/test_database.py
```python
import unittest
from BarterAI.src.database import Database

class TestDatabase(unittest.TestCase):

    @classmethod
    def setUpClass(cls):
        cls.db = Database()

    def test_connection(self):
        self.assertIsNotNone(self.db.connection, "Database connection failed.")

    def test_create_user(self):
        user_data = {
            "username": "test_user",
            "email": "test_user@example.com",
            "trade_history": [],
            "wishlist": []
        }
        user_id = self.db.create_user(user_data)
        self.assertIsNotNone(user_id, "User creation failed.")

    def test_get_user(self):
        user = self.db.get_user("test_user")
        self.assertIsNotNone(user, "User retrieval failed.")
        self.assertEqual(user["username"], "test_user", "User data incorrect.")

    def test_update_user(self):
        user_data = {
            "username": "test_user_updated",
            "email": "test_user_updated@example.com",
            "trade_history": [],
            "wishlist": []
        }
        result = self.db.update_user("test_user", user_data)
        self.assertTrue(result, "User update failed.")

    def test_delete_user(self):
        result = self.db.delete_user("test_user_updated")
        self.assertTrue(result, "User deletion failed.")

    def test_create_item(self):
        item_data = {
            "user_id": "test_user",
            "description": "test item",
            "photos": [],
            "desired_exchanges": []
        }
        item_id = self.db.create_item(item_data)
        self.assertIsNotNone(item_id, "Item creation failed.")

    def test_get_item(self):
        item = self.db.get_item("test_item")
        self.assertIsNotNone(item, "Item retrieval failed.")
        self.assertEqual(item["description"], "test item", "Item data incorrect.")

    def test_update_item(self):
        item_data = {
            "user_id": "test_user",
            "description": "test item updated",
            "photos": [],
            "desired_exchanges": []
        }
        result = self.db.update_item("test_item", item_data)
        self.assertTrue(result, "Item update failed.")

    def test_delete_item(self):
        result = self.db.delete_item("test_item_updated")
        self.assertTrue(result, "Item deletion failed.")

    @classmethod
    def tearDownClass(cls):
        cls.db.connection.close()

if __name__ == "__main__":
    unittest.main()
```